Move Mockserver fixture to higher conftest
