+++ testsuite/tests/kuadrant/authorino/authorization/opa/test_inline_rego.py
@@ -5,6 +5,12 @@ from testsuite.utils import rego_allow_header
 
 
 @pytest.fixture(scope="module")
+def header():
+    """Header used by OPA policy"""
+    return "opa", "opa-test"
+
+
+@pytest.fixture(scope="module")
 def authorization(authorization, header):
     """Adds OPA policy that accepts all requests that contain `header`"""
     authorization.add_opa_policy("opa", rego_allow_header(*header))
